+++ public/admin/_app/immutable/nodes/24.BR2KJCa9.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/tgDg0Jly.js";export{e as component};