+++ public/admin/_app/immutable/nodes/41.DAb2H5Y_.js
@@ -1,1 +1,0 @@
-import{It as e}from"../chunks/DFnvYPbU.js";import{c as t}from"../chunks/CHMRPlaG.js";import{t as n}from"../chunks/Bp2Urwao.js";var r=e({load:()=>i}),i=()=>{throw n(308,`${t}/settings/users`)};export{r as universal};